Backend/smart_india_hackathon/requirements.txt
asgiref==3.5.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==4.0.1
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
docopt==0.6.2
docutils==0.17.1
drf-yasg==1.20.0
fett==0.3.2
h11==0.12.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
networkx==2.6.3
outcome==1.1.0
packaging==21.0
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.3
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jsonrpc-server==0.3.4
pytz==2021.3
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
snooty-lextudio==1.11.6
sortedcontainers==2.4.0
sqlparse==0.4.2
tenacity==6.3.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
watchdog==1.0.2
wsproto==1.0.0
